. added btinrtst test, which is a version of helloworld with the built-in router. the advantage is that the traffic between the dbus is removed and the latency of an echo request can improve 10%.
. propids.h, frmwrk.h, prxyfdo.cpp, reqfwdr.cpp, rpcroute.cpp, dbusprxy.cpp,
prxyfdo.cpp, stmpdo.cpp, tcpfido.cpp, tcpport.cpp, bridge.cpp, ifstat.cpp: added
GetRouterName and SetRouterName to CIoManager which is to provide unique name of
the router across all the objects within the process, as was set by CIoManager's
GetModName or MODNAME_RPCROUTER, cannot work in the built-in-router environment.

. iftasks.cpp: fixed a bug in CIfTaskGroup::OnCancel, which did not test the
task queue before referencing the element in the queue and caused segment fault,
in the call of ClearActiveTasks when the CRpcReqForwarderProxyImpl is about to
stop.

. rpcroute.cpp: fixed a bug from CRpcRouter's OnPreStop, which was a
time-consuming operation involving long wait for interface objects to stop, and
caused deadlock since the tasks spawned are blocked on the task thread the
OnPreStop is waiting on. The solution at present is to run OnPreStop on a
stand-alone thread.

. ifhelper.h, rpcroute.h, bridge.cpp, dbusport.cpp, sfcli.cpp, rpcif.cpp,
dbushook.cpp: cleared some compile warnings in the release build.

. clsids.h, fdodrv.cpp, dbusport.cpp, dbusport.h, dbusprxy.cpp prxyport.h: added
CDBusProxyPdoLpbk and related support in the CDBusBusPort. It is added for the
builtin router support.  It will replace the CDBusProxyPdo in the builtin-router
application.