. rpcif.cpp: fixed a bug to allow the proxy to connect to port 443. Previously the port to connect was limited to be greater than 1024.
. rpcroute.cpp: fixed a bug that CRpcRouter's GetBridgeProxy returned wrong
error code other than -ENOENT, which caused the proxy failed to create.

. tcportex.cpp: fixed a bug that the `stream open response' could randomly
return ahead of the irp moving to the correct waiting queue, and caused the
stream object was not created and sfclitst failed immediately on opening the
stream channel.

. tcportex.cpp: fixed a potential race condition when more than stream creation
requests are being processed simutaneously, by changing the stream-id-counter to
an atomic object.

. driver.json, sfdesc.json, stmdesc.json: adjusted the connection parameters.