. objbase.cpp, defines.h: added GetModulePath and GetLibPath to get the path for library searching.
. clsids.cpp, drvmgr.cpp, frmwrk.h, iomgr.cpp, rpcif.cpp, propids.h: added
TryLoadClassFactory and TryFindDescFile to CIoManager to find the specified file
from a set of paths to search. Also the set search paths is added to the
CIoManager's registry.

. reqfwdr.cpp, rpcroute.cpp, k5server.cpp, security.cpp: Set an iomgr pointer to
the configdb to pass to the LoadObjDesc, so that the config file and class
factories can be found and loaded from a wider range of search paths.

. security.cpp: made the service name specified from command line to supercede
the one read from the config file by the CAuthentServer.