. Makefile.init.pub.in: made the change to obtain the include path and lib path via the pkg-config for dbus, glib, jsoncpp, which has better cross-platform support.
. dbusport.h: added to the CConnParamsProxy's less to check the difference of
propSSL if all the other checks are passed.

. security.cpp: mark the pStartIfTask and the pLoginTask to always call the
client-notify when the task is done, whethere synchronously or asynchronously,
from the BuildStartAuthProxyTask. it could fall into endless wait if the
CK5AuthProxy failed to start, due to the dbusproxy port is already disconnected.

. dbusprxy.cpp, k5proxy.cpp, k5proxy.h: made the dbusproxy pdo to send out a
eventConnErr if the eventRmtSvrOffline is received, instead of send out another
eventRmtSvrOffline to the pnpmgr's connection point. And accordingly, the
CK5AuthProxy is added OnEvent to handle the eventConnErr.