. sslfido.cpp stmpdo2.cpp: added irp unlocking to make sure CIoManager's CompleteIrp is run on a lock free context. Since if irp have CompleteInPlace set, the lock could affect the irp's callback to run properly which assume the irp is unlocked.
. bridge.cpp, multihop.cpp, reqfwdr.cpp, rpcroute.cpp, rpcroute.h, security.cpp,
security.h: replaced GetRouterName with propSvrInstName for all the router
objects. It could be different from the `module name', which contains no
instance information.

. security.h: rewrote the CheckReqToFwrd of CRpcRouterReqFwdrAuth. Since
GetRouterName was not a valid piece of information to build the objpath, which
consisted of the router name plus the object name, can now only be decided with
the informantion from bridge proxy. The reason is that the reqfwdr does not know
the peer router's name. And the ObjPath is now built with the bridge proxy's
propObjPath with the objname replaced from OBJNAME_TCP_BRIDGE_AUTH to
OBJNAME_ROUTER_BRIDGE_AUTH.

. btinrtst.cpp, btinrt.json, routmain.cpp: made changes plus the above changes
to get btinrtcli to work with authentication enabled. But btinrtsvr does not
work still, because it does not seem to be as useful as the client side. The
significance of btinrtcli is that the router names can be different between the
bridge and proxy, and there can be multiple instances of reqfwdr from the client
side.