. clsids.h, dbuserr.h, defines.h, streamex.h, utils.h, rpcfrmwrk.sip, mainloop.h, Makefile.init.pub.in: removed dependency of glib2. And gmloop.cpp and gmloop.h were removed from the project.
. gentoml.sh, pyproject.toml.tmpl: added the two files to determine the paths
for dbus-1's header files. and pyproject.toml was removed from the project.

. Makefiles: removed the redefinition of LINKER_OPTION from
test/Makefile.init.pub.in and make it refer to the LINKER_OPTION definition from
the master Makefile.init.pub.in, whose value is pkgconfig's output. And as a
result, the master LINKER_OPTION removed `-shared' option, which is then moved
to ipc and rpc's Makefile.in. And all the other makefiles were also changed due
to this fundamental change.