. ldbal.cpp, rpcroute.cpp, rpcroute.h, multihop.cpp, dbusprxy.cpp, ifstat.cpp, clsids.h, rpc/Makefile.in: added node backup/redundancy support.
. replaced ret = pTask->GetError() with ret = STATUS_PENDING for tasks scheduled
by RescheduleTask and AddSeqTask, if the two methods return success, because at
that point, the task is out of the control scope of current thread, and the
status code is not reliable. What the current thread can do is to return
STATUS_PENDING. Those occurances after AddAndRun and AppendAndRun are not
touched, because the task could be run either immediately or not, which need
some more investigations.