. ifhelper.h: continued to make the family functions of `CastTo' to work.
. iomgr.cpp, rpcif.cpp: added new path '/../etc/rpcf'to search for the object
desc files because the new layout of the installation directory structure.

. routmain.cpp, btinrtst.cpp, iftest.cpp: regression bug after introducing the
automake build system, that LoadObjDesc failed to search for the objdesc file if
it did not found the desc file at the input location, when the pIoMgr was not
passed to it.

. reqfwdr.cpp, security.cpp: regression bug introduced by the load-balance
support, that the propRouterPath was missed to return in the response if the
connection has existed already.