. config.cpp, configdb.h: optimized some key method in CCfgOpener. There should have 1ms performance improvement.
. defines.h: fixed a bug that ERROR_QUEUE_FULL and ERROR_NOT_IMPL are equal in
value.

. counters.cpp, counters.h, propids.h, bridge.cpp: added propWndSize and
refacted the CStatCountersServer for better performance.

. security.cpp, security.h: added error code to OnLoginFailed's input parameters
so that it can report accurate reason of the login failure.

. iftasks.cpp, rpcif.cpp: optimized CIfInvokeMethodTask's OnTaskComplete again.
and mainly brought back the FilterMessage from CRpcServices's SendResponse to
filter requests either with or without response, rather than with response only.

. asyncsvr.cpp, kasvr.cpp: fixed a bug that kasvrtst cannot reset timer on the
keepalive request from kaclitst, because of missing to call CInterfaceServer's
InitUserFuncs