. k5proxy.cpp, k5server.cpp, rpcroute.h, rpcroute.cpp, bridge.cpp: moved AppendConnParams and gen_sess_hash to rpcroute.cpp. And added session-hash to the non-authenticated sessions. And now the session-hash is prepended with 'AU' for authenticated sessions, and 'NA' for non-authenticated sessions.
. iftasks.cpp, iftasks.h: splitted CancelTaskChain into CancelTaskChain and
DoCancelTaskChain to allow GetFwrdTask to be accessed with task-lock.

. iftasks.cpp: modified the taskgroup's canceling mechanism. The canceling flat
only be set if OnCancel is called and cleared after it quits, unlike the
previous implementation, which could result in both OnCancel and RunTask quit
with tasks waiting in the queue.

. iftasks.cpp, rpcroute.h, rpcroute.cpp: moved CIfParallelTaskGrpRfc from
iftasks.h to rpcroute.h, and added ClearPendingsXXX methods for queued request
canceling when request-based flow-control enabled.

. dbusprxy.cpp, prxyport.h, reqfwdr.cpp, bridge.cpp: added propNoReply to the
reqCtx, so that, without deserializing the request message, the rpcrouter can
know whether to send a response or not when canceling a pending request.

. bridge.cpp, reqfwdr.cpp: added propPath2 to the reqCtx, which will be passed
on to the server app rather than propRouterPath, which will finally reduce to '/'
and useless for the server app.

. msgmatch.h: replacing CCfgOpenerObj with CCfgOpener as much as possible.