. iftasks.cpp: fixed a segment fault in GetEndFwrdTask, which discarded the reference to the task second the the last one, and caused the pointer to the task to return became invalid.
. bridge.cpp, reqfwdr.cpp: fixed a bug in the RequeueInvTask, that missing to
mark the task pending to requeue caused the message unable to be requeued and
lost permernantly.

. rpcroute.cpp: added MarkPending to CIfParallelTaskGrpRfc's AddAndRun if the
task is still in pending state. It is possible the task is still in the pending
queue of the task group after the call to AddAndRun, and the task will left
without pending marked without this MarkPending call.