. Helpers.java, JRetVal.java, proxy.i, rpcbase.i, JavaRpcService.java: Java support development continuing...
. ifhelper.h, iftasks.cpp, proxy.sip, gencpp.cpp: made change to
DEFER_CANCEL_HANDLER2 to install the cancel callback after the
CIfInvokeMethodTask's completion is done, as was installed before the
CIfInvokeMethodTask completion occurs, which break the contract that user's
request handler will have a pointer to the CIfInvokeMethodTask, which could add
complexity when the invoke task be chained by other tasks.

. proxy.sip: small optimization in OnStmClosing, to avoid release the python
object twice.

. proxy.sip, proxy.py, server.sip, genpy.cpp: renamed InstallCompNotify to
InstallCancelNotify.

. gencpp.cpp, genpy.cpp, reqfwdr.cpp: fixed some minor bugs.

. ridl.7, ridlc.y, astnode.h, ridl/README.md: added check of duplicated method
names within a interface declaration, and throwing error when duplicated names
occurs..