. evtest.ridl, EventTestcli.java evtest/maincli.java, evtest/mainsvr.java: added testcase 'evtest'.
. genjava.cpp: fixed a bug in generating the maincli code for event-only
services, that the comment mistakenly calls the event handler to send a request.

. JavaRpcService.java: fixed a bug in filtering event request, which compare the
incoming request name with the wrong prefix 'UserMethod_", and failed to invoke
event handler constantly.

. seribase.cpp, seribase.h, gencpp.cpp: made a change the ridl serialization
code for C++ code generator, to support HSTREAM serialization in all the
possible situations as Python and Java does. Previously, the HSTREAM
serialization was restricted to the formal args in the request only. And now,
the HSTREAM serialization can happen as a struct field, array element, map
element. It also fixed some wrong code around the HSTREAM
serialization/deserialization.