. configure.ac, Makefile.am, java/Makefile.am, examples/Makefile.am, examples directory: added java and examples directories to the main build script.
. genjava.cpp: added method getDescPath to both maincli and mainsvr classes, so
that, the application will not fail if the working directory is not the
directory the class files reside.

. serial.cpp, stlcont.h: fixed a bug that CStlQwordVector did not implement
Deserialize( const char*, guint32 ), and CConfigDb tried to deserialize it and
failed with error -95.

. rpcroute.cpp: fixed a bug in CRouterStopBridgeTask::RunTask, that the
DoSyncedTask happens at the same time the cancel requests are sending to the
servers, and CRpcReqForwarderProxy turned to stateStopping and failed to send
the requests.

. bridge.cpp: fixed a bug in RetrieveTaskId, which mistakenly used the
keep-alive taskid as the task-id to cancel on the remote server, and thus failed
to cancel the desired request because the server cannot find a task with the
wrong task id.