. fusedefs.h, fuseif.cpp: replaced FIONREAD with FIOC_GETSIZE because the ioctl with the former one cannot reach fuseop_ioctl and fs_ioctl of CFuseStmFile.
. fuseif.cpp, fuseif.h: changed the fs_write_buf of CFuseStmFile to use
WriteStream instead of WriteStreamNoWait, because the latter one has size limit
and failed in large data transfer.

. fuseif.cpp, fuseif.h, genfuse.h, gencpp.cpp: overrode the
OnStmRecv from CStreamSyncBase to generate poll notification on data arrival,
which is important for polling users.

. server.sip: fixed a bug that GetIdHashByChan did not return the response in
the expected list, but a handle value of int64, which caused exception in
SerialHStream from seribase.py.

. genfuse.cpp, serijson.cpp, serijson.h: fixed a bug in Serializing HSTREAM for
fuse due to the request change of HSTREAM value from a INT64 value to a string
of file name.