. fuse_i.h, fusedefs.h, fuseif.cpp, fuseif.h, fuseif_ll.cpp: added support of multiple instances of services by making dir of more service directories.
. iftasks.cpp: fixed a bug that FindTask could cause deadlock when searching
subgroups without lock released.

. rpcif.cpp: fixed a bug in server side SendResponse, that closed fds blindly
without checking if it is a stream id or a pipe fd.

. streammh.cpp, streammh.h stmrelay.h, stmrelay.cpp: fixed a potential bug when
wrapping IStream::OnConnected with the DEFER_IFCALLEX_NOSCHED, which failed the
dynamic_cast from CStreamServerRelay to IStream, and the OnConnected
notification failed to send. It is potential bug because the OnConnected handler
does nothing so far.

. stmrelay.cpp: fixed a bug in CIfStartUxSockStmTask's OnTaskComplete, that
there was a chance the response property was not set and the FetchData response
may build a message with an invalid file descriptor and cause the reqfwdr to fail
deserialization of the message.

. iolib.py, dumpStream.py, testypes/maincli.py, testypes/mainsvr.py: added some
test scripts