. fuseif.cpp: changed the content formant of svcstat and conn_params to Json strings.
. fuseif.cpp: made connection directory removable when the number of `svc point`
under the directory drops to zero.

. fusedefs.h, fuseif.h, fuseif.cpp: added file 'clslist' to list the available
services to start.

. fuseif.h, fuseif.cpp: fixed a bug in AddSvcPoint which start the proxy
interface in the parallel, which has the race condition when the port is the
same for the two proxies, and one of the proxies could submit irp assuming the
port was ready, but actually still in startup and failed both proxies by
shutting down the connection to the remote server.