. fuseif.cpp, fuseif.h: fixed a bug that 'reload' and 'restart' cannot work under multihop environment. The solution is to add a FindSvcDir to CFuseDirectory to find the specified 'service point' recursively through all the sub-directiories, which was not supported in the old solution.
. bridge.cpp, k5proxy.cpp, k5server.cpp, security.cpp: fixed a bug that the
client and the server have different session hash, due to the client was behind
a NAT gateway, and either had different source ip address and therefore,
generated different session hashs. The solution is that the server dispatch its
session hash to the proxy upon login, and so both party have the same session
hash.