. fastrpc.h: fixed a minor bug in OnStartStmComplete from CRpcStmChanBase, that SetPreStopStep should only be called when the irp is a 'stop' irp.
. irpthread.cpp: removed the code snippet that scanning task queue for duplicate
task in CTaskQueue's AddTask, which is completely not necessary in 99% of all
the cases.

. streamex.cpp: fixed a memory leak because timedout worker irp has the
posibility to be completed by the timer and the irp from the uxport at the same
time, and now we set the timer has the right to complete the irp, because the
uxport's attempt to complete irp could possibly fail because the irp's state is
already set to IRP_STATE_CANCELLED, while the timer's completion as goes to
OnWorkerIrpComplete has special processing on this condition, and can complete
the irp successfully all the time.

. port.cpp: fixed a bug in OpenPdoPort from CGenericBusPort, that the bus port's
state can change frequently  between PORT_STATE_READY and
PORT_STATE_BUSY_SHARED, and the bug failed OpenPdoPort when
the bus port has switched to PORT_STATE_BUSY_SHARED for ClosePdoPort, which has
no risk for OpenPdoPort to move on. So the fix enabled OpenPdoPort to run when
the bus port is under state PORT_STATE_BUSY_SHARED.