. dbusport.h, rpcpdo.cpp, tcpfido.cpp, tcportex.cpp: fixed a bug that on disconnection, the CReqFwdrOpenRmtPortTask cannot complete immediately because the IRP in the CRpcTcpFido, and CRpcNativeProtoFdo was not canceled immediately. the solution is to add CancelAllIrps once the disconnection is detected.
. iftasks.cpp: fixed a bug that if the interface object failed during
CIfOpenPortTask, it did not change to stateStartFailed due to the port has not
registered with the 'IO manager'. The solution is to call SetStateOnEvent
explicitly if the port has not been registered with the 'IO mangager'.

. ifhelper.h, iftasks.cpp, pnpmgr.cpp: rewrote the error handler of
BuildPortStack, as the earlier one would leave the tcp connection alive while
the port object and the proxy object have stopped and been removed. The new
solution is to let the partialy built port stack to complete start and then
follows the normal port shutdown process, which can shutdown the socket
gracefully.

reqfwdr.cpp, rpcroute.cpp, rpcroute.h: fixed a bug that the bridge proxy could
last till CoUninitialize, outlive its parent object, because the previous
shutting down approach was to leav the current synchronization context, and
add a new 'shutdown task' behind the tail task of the CRpcRouterReqFwdr. and
the new 'shutdown task' was then left pending while the CRpcRouterReqFwdr has
quit. The solution is added a StopIfSafe method, which guarantees to shutdown
the bridge/bridge proxy object within the same synchronization context.