. wscfg2.json.in: fixed a bug, that currently gmssl does not support connection with a websock server because webserver does not talk in gmssl. And the gmssl test has to disable the websock connection.
. gmsslkey.sh, opensslkey.sh, rpcfg.py, updinitcfg.py: added 'secret file'
installation to the key installation process. And added SSL type flag( openssl
or gmssl ) to the installer's name. And for non-SSL connection and SSL with demo
key generation, the rpcfg.py will generate both server and client installers for
deployment. BTW, the installer name for non-SSL connection does not has SSL type
flag.