. ifstat.cpp: fixed a memory leak by changing the state transition from stateStarting-stateStopped on cmdShutdown to stateStarting-stateStopping on cmdShutdown, because the partial started interface object requires a full stop process to avoid some resources not released.
. pnpmgr.cpp: fixed a memory leak that if the CIfOpenPortTask was canceled right
before the pnp manager notify it the port-object successfully created, the port
object could resides in the memory without access any more.

. iftasks.cpp, iftasks.h: added OnCancel to CIfOpenPortTask for debug purpose.
and made a change to put the CIfTaskGroup to stateStopping immediately after its
task queue is empty, as different from the earlier design which set
stateStopping when it entered OnComplete, that is, there was a slim time window
the taskgroup can accept tasks and then quit and discard those tasks. And also
made a change to CIfDummyTask to set its return code if its OnTaskComplete or
OnIrpComplete is called.